Add test for invalid package name
We run validation for package names but it turns out this wasn't being
tested and had a bug. The `validate` function returns an object with
results, but we were treating it like a boolean, so it was always
passing.

The test adds an invalid package name (`http`) and makes sure the error
bubbles.